презентация 12.12.14

+++ app/platforms/ios/www/js/controllers/cart.js
@@ -270,23 +270,41 @@ BauVoiceApp.controller('CartCtrl', ['$scope', 'localDB', 'localStorage', '$locat
 
 
   //----- Delete Product
-  $scope.deleteProduct = function(productIdBD, productIdArr) {
-    $scope.global.productObj.splice(productIdArr, 1);
-    $scope.cart.productObjSource.splice(productIdArr, 1);
-    $scope.cart.allAddElements.splice(productIdArr, 1);
-    --$scope.global.productCounter;
-    // Change order price
-    $scope.global.calculateOrderPrice();
-    localDB.deleteDB($scope.global.productsTableBD, {'orderId': {"value": $scope.global.orderNumber, "union": 'AND'}, "productId": productIdBD});
-    localDB.deleteDB($scope.global.componentsTableBD, {'orderId': {"value": $scope.global.orderNumber, "union": 'AND'}, "productId": productIdBD});
-    localDB.deleteDB($scope.global.visorsTableBD, {'orderId': {"value": $scope.global.orderNumber, "union": 'AND'}, "productId": productIdBD});
-    localDB.deleteDB($scope.global.windowSillsTableBD, {'orderId': {"value": $scope.global.orderNumber, "union": 'AND'}, "productId": productIdBD});
-    //----- if all products were deleted go to main page????
-    if(!$scope.global.productCounter) {
-      $scope.global.orderTotalPrice = 0;
-      $scope.global.calculateOrderPrice();
+  $scope.clickDeleteProduct = function(productIdBD, productIdArr) {
+
+    navigator.notification.confirm(
+      'Хотите удалить продукт?',
+      deleteProduct,
+      'Удаление!',
+      ['Да','Нет']
+    );
+
+    //----- Delete Product
+    function deleteProduct(button) {
+      if(button == 1) {
+
+                                    $scope.global.productObj.splice(productIdArr, 1);
+                                    $scope.cart.productObjSource.splice(productIdArr, 1);
+                                    $scope.cart.allAddElements.splice(productIdArr, 1);
+                                    --$scope.global.productCounter;
+                                    localDB.deleteDB($scope.global.productsTableBD, {'orderId': {"value": $scope.global.orderNumber, "union": 'AND'}, "productId": productIdBD});
+                                    localDB.deleteDB($scope.global.componentsTableBD, {'orderId': {"value": $scope.global.orderNumber, "union": 'AND'}, "productId": productIdBD});
+                                    localDB.deleteDB($scope.global.visorsTableBD, {'orderId': {"value": $scope.global.orderNumber, "union": 'AND'}, "productId": productIdBD});
+                                    localDB.deleteDB($scope.global.windowSillsTableBD, {'orderId': {"value": $scope.global.orderNumber, "union": 'AND'}, "productId": productIdBD});
+                                    //----- if all products were deleted go to main page????
+                                    if($scope.global.productCounter > 0 ) {
+                                    // Change order price
+    
+                                    $scope.global.calculateOrderPrice();
+                                    } else {
+                                    $scope.global.createNewProjectCart();
+                                    }
+
+      }
+
     }
   };
+
 
 
 
@@ -305,7 +323,9 @@ BauVoiceApp.controller('CartCtrl', ['$scope', 'localDB', 'localStorage', '$locat
   $scope.lessProduct = function(productIdBD, productIdArr) {
     newProductsQty = $scope.global.productObj[productIdArr].productQty;
     if(newProductsQty === 1) {
-      $scope.deleteProduct(productIdBD, productIdArr);
+      //$scope.deleteProduct(productIdBD, productIdArr);
+      $scope.clickDeleteProduct(productIdBD, productIdArr);
+
     } else {
       $scope.global.productObj[productIdArr].productQty = --newProductsQty;
       //oldProductPrice = parseFloat($scope.cart.productObjSource[productIdArr].productPrice);
@@ -375,6 +395,8 @@ BauVoiceApp.controller('CartCtrl', ['$scope', 'localDB', 'localStorage', '$locat
     $scope.global.productEditNumber = false;
     $scope.global.templateIndex = 0;
     $scope.global.profileIndex = 0;
+    $scope.global.product = angular.copy($scope.global.productSource);
+    $scope.global.clearAllAddElements();
     $scope.global.prepareMainPage();
     $location.path('/main');
   };
